int to size_t and corrections

+++ xmldb_lib/StringList.h
@@ -6,18 +6,18 @@
 #include "string.h"
 #include "String.h"
 struct StringList{
-    int length;
-    int size;
+    size_t length;
+    size_t size;
     struct String* items;
 };
 
-void init_StringList(struct StringList*v,int size);
-void StringList_Resize(struct StringList *vect,int new_size);
+void init_StringList(struct StringList*v,size_t size);
+void StringList_Resize(struct StringList *vect,size_t new_size);
 void free_StringList(struct StringList *vect);
 void appendto_StringList(struct StringList *src_dest, struct String* string);
 void concatenate_StringList(struct StringList *dest, struct StringList* src);
-void insertInTo_StringList(struct StringList *src_dest, int index, struct String* string);
-void removeFrom_StringList(struct StringList *src_dest, int index);
+void insertInTo_StringList(struct StringList *src_dest, size_t index, struct String* string);
+void removeFrom_StringList(struct StringList *src_dest, size_t index);
 void String_Split(struct StringList *result ,struct String *string, char* Separator);
 void clear_StringList(struct StringList *vect);
 void free_StringListReturn(struct StringList *vect);
